ensure that % is escaped during percent encoding
Since the % sign is used for the escape in percent encoding, we need to
ensure that that is itself escaped if found in the input. Debian git tags
in vcs_urls in particular seem to be moving to replacing the : in a
version like `1:41.1` into `1%41.1`. The correct encoding of that winds
up being `1%2541.1`. Without this, adding the vcs_url as it is given
will wind up forcing the _parsing_ of the subsequent displayed
PackageUrl to decode the `1%41.1` as `1A.1`.

Fixes #12.